+++ src/share.rs
@@ -16,7 +16,6 @@ use crate::{Message, Source};
 ///
 /// See <https://github.com/staltz/callbag-share/blob/d96748edec631800ec5e606018f519ccaeb8f766/index.js#L1-L32>
 ///
-///
 /// [RxJS `share`]: https://rxjs.dev/api/operators/share
 pub fn share<T: 'static, S>(source: S) -> Source<T>
 where